+++ qcommon/src/main/java/jake2/qcommon/network/messages/server/DownloadMessage.java
@@ -12,13 +12,13 @@ import java.util.Arrays;
 public class DownloadMessage extends ServerMessage {
 
     public byte[] data;
-    public int percentage;
+    public byte percentage;
 
     public DownloadMessage() {
         super(ServerMessageType.svc_download);
     }
 
-    public DownloadMessage(byte[] data, int percentage) {
+    public DownloadMessage(byte[] data, byte percentage) {
         this();
         this.data = data;
         this.percentage = percentage;
@@ -32,14 +32,14 @@ public class DownloadMessage extends ServerMessage {
             SZ.Write(buffer, data, data.length);
         } else {
             MSG.WriteShort(buffer, -1);
-            MSG.WriteByte(buffer, 0);
+            MSG.WriteByte(buffer, (byte) 0);
         }
     }
 
     @Override
     public void parse(sizebuf_t buffer) {
         int size = MSG.ReadShort(buffer);
-        percentage = MSG.ReadByte(buffer);
+        percentage = (byte) MSG.ReadByte(buffer);
         if (size != -1) {
             data = new byte[size];
             MSG.ReadData(buffer, data, size);
